Refactor: Organize Activities into `activity` package
This commit moves `MainActivity` and `HistoryActivity` into a new `activity` subpackage to improve project structure.

Key changes:
- Moved `MainActivity.kt` to `activity/MainActivity.kt` and updated its package declaration.
- Moved `HistoryActivity.kt` to `activity/HistoryActivity.kt` and updated its package declaration.
- Updated `AndroidManifest.xml` to reflect the new activity paths for `MainActivity`, `MainActivityAlias`, and `HistoryActivity`.
- Changed `android:label` for `MainActivityAlias` to `@string/app_name`.
- Set `android:enabled="false"` for `MainActivityAlias` to ensure it's not the default launcher activity initially.
- Updated `HistoryTileService.kt` to import `HistoryActivity` from the new `activity` package.
- Removed unused imports from `MainActivity.kt`.
