+++ app/src/main/AndroidManifest.xml
@@ -21,17 +21,16 @@
         android:theme="@style/Theme.LiveInPeace"
         tools:ignore="UnusedAttribute">
         <activity
-            android:name=".MainActivity"
-            android:exported="true"
-            android:label="@string/title_activity_main"
+            android:name=".activity.MainActivity"
             android:theme="@style/Theme.LiveInPeace">
         </activity>
         <activity-alias
             android:name=".MainActivityAlias"
             android:exported="true"
-            android:label="@string/title_activity_main"
+            android:label="@string/app_name"
             android:theme="@style/Theme.LiveInPeace"
-            android:targetActivity=".MainActivity">
+            android:enabled="false"
+            android:targetActivity=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -39,7 +38,7 @@
             </intent-filter>
         </activity-alias>
         <activity
-            android:name=".HistoryActivity"
+            android:name=".activity.HistoryActivity"
             android:theme="@style/Theme.LiveInPeace" />
 
         <service
