+++ app/src/main/java/com/maary/liveinpeace/activity/MainActivity.kt
@@ -1,16 +1,9 @@
-package com.maary.liveinpeace
+package com.maary.liveinpeace.activity
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.maary.liveinpeace.ui.screen.SettingsScreen
 import com.maary.liveinpeace.ui.theme.LiveInPeaceTheme
 import dagger.hilt.android.AndroidEntryPoint
