+++ app/src/main/java/com/maary/liveinpeace/service/HistoryTileService.kt
@@ -5,7 +5,7 @@ import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import android.os.Build
 import android.service.quicksettings.TileService
-import com.maary.liveinpeace.HistoryActivity
+import com.maary.liveinpeace.activity.HistoryActivity
 
 class HistoryTileService: TileService() {
     override fun onClick() {
